Create Day25
Day 25 of #dailycoding challenges ⬇️
After working on the decision trees, today we are going to work on random forest classifier, so we are going to work on the same database (readingSkills) of the library "party". and we will use this time the library "randomForest"
Random forest Classifier explication:
Random forest, like its name implies, consists of a large number of individual decision trees that operate as an ensemble. Each individual tree in the random forest spits out a class prediction and the class with the most votes becomes our model’s prediction .

Take a look at the code it's really amazing ⬇️
next time we will work on randomForest with the library "caret"

Happy Coding Learning !

#dailycoding #challenge #programming #rstats #R #rladies